+++ client/sports-course/src/lib/graphql.ts
@@ -56,7 +56,6 @@ const errorLink = onError(({ graphQLErrors, networkError, operation, forward }) 
                     getNewToken()
                         .then(({ data }: any) => {
                             const { refreshToken } = data
-                            console.log("errorLink ", refreshToken?.token)
                             useStore.setState({ isAuth: true, token: refreshToken?.token });
                             resolvePendingRequests();
                             return true;
@@ -84,7 +83,9 @@ const errorLink = onError(({ graphQLErrors, networkError, operation, forward }) 
         } else {
             //console.log(`[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`);
         }
-        console.log(`[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`);
+        if (process.env.NODE_ENV !== "production") {
+            console.warn(`[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`);
+        }
     }
 });