+++ client/team/lib/graphql.ts
@@ -41,7 +41,6 @@ const errorLink = onError(({ graphQLErrors, networkError, operation, forward }) 
                     getNewToken()
                         .then(({data}: any) => {
                             const { refreshToken } = data
-                            console.log("errorLink ", refreshToken?.token)
                             useStore.setState({ isAuth: true, token: refreshToken?.token });
                             resolvePendingRequests();
                             return true;
@@ -67,7 +66,9 @@ const errorLink = onError(({ graphQLErrors, networkError, operation, forward }) 
                 return forward(operation);
             });
         } else {
-            console.log(`[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`);
+            if (process.env.NODE_ENV !== "production") {
+                console.warn(`[GraphQL error]: Message: ${message}, Location: ${locations}, Path: ${path}`);
+            }
         }
     }
 });